:hammer: refactor(model): split mesh and model
