+++ .vscode/settings.json
@@ -6,6 +6,7 @@
     "rust-analyzer.linkedProjects": [
         ".\\Cargo.toml",
         ".\\Cargo.toml",
+        ".\\Cargo.toml",
         ".\\Cargo.toml"
     ],
     "code-runner.executorMap": {
