+++ Cargo.toml
@@ -13,7 +13,6 @@ thiserror = "1" # error derive macro
 anyhow = "1" # error handling
 image = "0.24.5" # image processing
 nalgebra = "0.32.2" # Linear algebra : https://nalgebra.org/docs/
-russimp = { version = "=2.0.5", features=["prebuilt"] } # model loading
 tobj = "4.0.0" # OBJ loader
 
 # window related dependencies
