+++ src/lib.rs
@@ -2,6 +2,7 @@ mod buffer;
 mod camera;
 mod light;
 mod material_phong;
+mod mesh;
 mod model;
 mod shader;
 mod texture;
@@ -14,6 +15,7 @@ pub use buffer::*;
 pub use camera::*;
 pub use light::*;
 pub use material_phong::*;
+pub use mesh::*;
 pub use model::*;
 pub use shader::*;
 pub use texture::*;
